pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "users"
version = "0.1.0"
description = "A flexible user management package for FastAPI backends"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "pydantic>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.5",
    "sqlalchemy>=2.0.0",
    "alembic>=1.8.0"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
]

[project.optional-dependencies]
mongodb = ["motor>=3.0.0", "pymongo>=4.0.0"]
redis = ["redis>=4.0.0"]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.20.0", "black>=22.0.0", "isort>=5.0.0", "mypy>=1.0.0"]

[project.urls]
Homepage = "https://github.com/yourusername/users"
Repository = "https://github.com/yourusername/users"

[tool.setuptools.packages.find]
where = ["."]
include = ["users*"]